% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{glance.LDA}
\alias{glance.LDA}
\title{Implementation of broom::glance() for LDA objects}
\usage{
glance.LDA(L)
}
\arguments{
\item{L}{an LDA object, as produced by LDA()}
}
\value{
a data.frame of eigenvalues from the associated generalized eigenvalue problem
}
\description{
Produces a one-line summary of an LDA object
}
\examples{
\dontrun{
model <- LDA( Species ~ ., iris )
broom::glance( model )
##       VQ1       VQ2
## 1 23.42386 0.2167796
##
}
}
