% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{gsep}
\alias{gsep}
\title{Generalized Symmetric Eigenvalue Problem (GSEP) Solver}
\usage{
gsep(A, B, nz = nrow(A))
}
\arguments{
\item{A}{n-by-n matrix A.}

\item{B}{n-by-n matrix B.}

\item{nz}{integer indicating the number of eigenvalues /-vectors to obtain. Defaults to n.}
}
\value{
A list with two elements. $d contains the eignavalues in decreasing order. $v contains the corresponding eigenvectors.
}
\description{
Finds the first nz eigenvalues/-vectors of the generalized
  eigenvalue problem \eqn{A v = \lambda B v}, where A and B are assumed to be symmetric.
}
