% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.LDA}
\alias{tidy.LDA}
\title{Implementation of broom::tidy() for LDA objects}
\usage{
tidy.LDA(L)
}
\arguments{
\item{L}{an LDA object, as produced by LDA()}
}
\value{
a data.frame with LDA component loadings
}
\description{
Produces a summary of an LDA object, expressed as a data.frame
}
