% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.LDA}
\alias{tidy.LDA}
\title{Implementation of broom::tidy() for LDA objects}
\usage{
tidy.LDA(L)
}
\arguments{
\item{L}{an LDA object, as produced by LDA()}
}
\value{
a data.frame with LDA component loadings
}
\description{
Produces a summary of an LDA object, expressed as a data.frame
}
\examples{
\dontrun{
model <- LDA( Species ~ ., iris )
broom::tidy( model )
##       Feature      LDA1        LDA2
## 1 Sepal.Length  0.626970 -0.01449191
## 2  Sepal.Width  1.181041 -2.14917074
## 3 Petal.Length -2.135565  0.52600987
## 4  Petal.Width -1.828151 -1.89227871
##  
}
}
